+++ NEWS.md
@@ -1,3 +1,13 @@
+# brickr 0.3.4
+
+* Error fixes for dplyr 1.0.0 release.
+
+## Mosaics
+
+* When uploading a png with a transparent background, those bricks will default to White. Change input 'trans_bg' to any LEGO color name.
+
+----
+
 # brickr 0.3.2
 
 * An overall leaner package to ensure optimal performance and remove experimental features.